fix: error for empty weight
The error was that sometimes the workout history didn't store the weight
of some exercises. After some debugging, I realized that it was due to
the fact that I forgot to press "Start" before actually putting the
data. Basically it didn't record the history for the data that was
entered before pressing the "Start" button.

Solution: adding a modal here to confirm before starting the workout so
we have initialdata in all cases (there is still one check left). But
for now a confirmation dialog is there.